vim - Add keybinding to close quickfix window
This is kind of an ugly hack, since the quickfix window is a buffer, we
can give it a local key binding and the problem is that timeoutlen is
the time that vim waits between keypresses for chained commands such as
`qa` or `<leader>w`. We temporarily set it to zero and then set it to
it's original value again. Hardcoded. Yeah.